requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
gxformat2
packaging
pydantic
requests
pyyaml
typing-extensions
urllib3